Place PushbackSideInputDoFnRunner in the correct directory
Fixes backport issue.
